死亡時にやり直せるように

+++ src/main/java/plugin/mininggamek/data/PlayerData.java
@@ -9,7 +9,7 @@ public class PlayerData {
 
   private String name;
   private int score;
-  private long gameTime;
+  private int gameTime;
   private double locationX;
   private double locationY;
   private double locationZ;
